requirements.txt
torch==1.11.0
torchvision==0.12.0
ninja==1.10.2
tqdm
imageio==2.9.0
imageio-ffmpeg
mrcfile
psutil
tensorboardX==2.5
pyspng
opencv-python
click
scipy
pyyaml
yacs
timm
termcolor
mtcnn
tensorflow==2.2.1
kornia
dominate
scikit-image
trimesh
numpy
protobuf==3.20.3